fix evm bridges to hyperliquid
