+++ src/views/Bridge/hooks/useBridgeAction/strategies/evm.ts
@@ -1,4 +1,4 @@
-import { BigNumber, constants, providers } from "ethers";
+import { BigNumber, constants, providers, utils } from "ethers";
 
 import {
   ChainId,
@@ -10,6 +10,10 @@ import {
   sendDepositTx,
   sendSwapAndBridgeTx,
   acrossPlusMulticallHandler,
+  fixedPointAdjustment,
+  generateHyperLiquidPayload,
+  getToken,
+  hyperLiquidBridge2Address,
 } from "utils";
 import { useConnectionEVM } from "hooks/useConnectionEVM";
 import { ERC20__factory } from "utils/typechain";
@@ -253,9 +257,11 @@ export class EVMBridgeActionStrategy extends AbstractBridgeActionStrategy {
       throw new Error("'transferQuote.quotedFees' is required");
     }
 
-    const message = await this.signHyperliquidMessage(params);
+    // For Hyperliquid, we need to switch to Arbitrum for permit signing
+    // then switch back to the source chain for the deposit
+    const message = await this._generateHyperliquidMessage(params);
 
-    // Ensure we are on the correct network
+    // Ensure we are on the correct network for the deposit
     await this.assertCorrectNetwork(selectedRoute.fromChain);
 
     const { spokePool } = await getSpokePoolAndVerifier(selectedRoute);
@@ -274,6 +280,95 @@ export class EVMBridgeActionStrategy extends AbstractBridgeActionStrategy {
     );
 
     return tx.hash;
+  }
+
+  /**
+   * Generate Hyperliquid message with permit signature
+   * This method handles the network switching for permit signing
+   */
+  private async _generateHyperliquidMessage(params: DepositActionParams) {
+    const { depositArgs, transferQuote, selectedRoute } = params;
+
+    if (!this.evmConnection.signer) {
+      throw new Error("'signer' is required");
+    }
+
+    if (selectedRoute.externalProjectId !== "hyperliquid") {
+      throw new Error(
+        "'selectedRoute.externalProjectId' must be 'hyperliquid'"
+      );
+    }
+
+    if (!transferQuote || !transferQuote.quotedFees) {
+      throw new Error(
+        "'transferQuote' and 'transferQuote.quotedFees' are required"
+      );
+    }
+
+    // Store current chain to restore later
+    const currentChainId = this.evmConnection.chainId;
+
+    // Switch to Arbitrum for permit signing (required for EIP-712 signature)
+    if (currentChainId !== ChainId.ARBITRUM) {
+      await this.evmConnection.setChain(ChainId.ARBITRUM);
+    }
+
+    try {
+      // Subtract the relayer fee pct just like we do for our output token amount
+      const amount = depositArgs.amount.sub(
+        depositArgs.amount
+          .mul(depositArgs.relayerFeePct)
+          .div(fixedPointAdjustment)
+      );
+
+      // Build the payload
+      const hyperLiquidPayload = await generateHyperLiquidPayload(
+        this.evmConnection.signer,
+        depositArgs.toAddress,
+        amount
+      );
+
+      // Create a txn calldata for transfering amount to recipient
+      const erc20Interface = new utils.Interface([
+        "function transfer(address to, uint256 amount) returns (bool)",
+      ]);
+
+      const transferCalldata = erc20Interface.encodeFunctionData("transfer", [
+        depositArgs.toAddress,
+        amount,
+      ]);
+
+      // Encode Instructions struct directly
+      const message = utils.defaultAbiCoder.encode(
+        [
+          "tuple(tuple(address target, bytes callData, uint256 value)[] calls, address fallbackRecipient)",
+        ],
+        [
+          {
+            calls: [
+              {
+                target: getToken("USDC").addresses![ChainId.ARBITRUM],
+                callData: transferCalldata,
+                value: 0,
+              },
+              {
+                target: hyperLiquidBridge2Address,
+                callData: hyperLiquidPayload,
+                value: 0,
+              },
+            ],
+            fallbackRecipient: depositArgs.toAddress,
+          },
+        ]
+      );
+
+      return message;
+    } finally {
+      // Restore original chain if it was different
+      if (currentChainId !== ChainId.ARBITRUM) {
+        await this.evmConnection.setChain(currentChainId);
+      }
+    }
   }
 
   private async _approveTokens(params: {
